temp_deploy_staging/generate-xml/requirements.txt
google-cloud-storage==2.18.2
google-cloud-firestore==2.16.0
google-cloud-pubsub==2.23.1
pandas==2.2.3
google-cloud-secret-manager==2.20.2
google-cloud-logging==3.10.0
